+++ client/src/components/settings/SettingContainer.jsx
@@ -1,0 +1,5 @@
+import { classNames } from '../../helpers/classnames';
+
+export const SettingContainer = ({ className, children }) => {
+    return <div className={classNames('setting', className)}>{children}</div>;
+};
