--- client/src/components/settings/SettingsContainer.js
@@ -1,4 +1,0 @@
-export const SettingsContainer = (props) => {
-    const className = props.className;
-    return <div className={className}>{props.children}</div>;
-};
