+++ client/src/index.scss
@@ -41,6 +41,7 @@ body {
 @import './styles/notification.scss';
 @import './styles/languageselector.scss';
 @import './styles/slider.scss';
+@import './styles/setting.scss';
 
 // Basic bigger components (umm this could use a better name?)
 @import './styles/footer.scss';
