+++ client/src/styles/avatarcreator.scss
@@ -1,8 +1,8 @@
 .avatar-creator-container {
-    padding: 0.2rem;
-    margin: 0.2rem;
-    background-color: $setting-background;
-    color: darken($app-text, 15%);
+    // padding: 0.2rem;
+    // margin: 0.2rem;
+    // background-color: $setting-background;
+    // color: darken($app-text, 15%);
 
     display: grid;
     grid-template-columns: 2em auto 2em;
