+++ client/src/styles/gamesettings.scss
@@ -149,56 +149,6 @@
     flex-direction: column;
 }
 
-.setting {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: nowrap;
-    justify-content: space-between;
-    align-items: center;
-
-    padding: 0.2rem 0.8rem;
-    margin: 0.2rem 0;
-    border-radius: 2px;
-    background-color: $setting-background;
-
-    &.text-control {
-        flex-wrap: wrap;
-    }
-
-    .icon-and-text {
-        align-items: center;
-        width: fit-content;
-        display: flex;
-        flex-grow: 0;
-        flex-shrink: 1;
-        margin-right: 1em;
-    }
-
-    .control {
-        display: flex;
-
-        flex-grow: 1;
-        margin: auto 0;
-        align-items: center;
-        justify-content: flex-end;
-
-        width: fit-content;
-
-        > .text-control {
-            flex-grow: 1;
-        }
-    }
-
-    &:hover {
-        background-color: $setting-background-hover;
-        color: $header-button-text-hover;
-
-        .control {
-            text-shadow: none;
-        }
-    }
-}
-
 .imported-card-packs {
     display: flex;
     flex-direction: column;
